skills/ppt-knowledge-doc/scripts/requirements.txt
python-pptx==1.0.2
pydantic==2.12.5
pillow==12.1.0
pytesseract==0.3.13
